Created a class for managing adding different timed effects to a world (#823)
added an effect to multiply glucose spawn density by 0.8 every editor session

+++ src/microbe_stage/patch_manager.h
@@ -12,7 +12,7 @@ class PatchManager : public Leviathan::PerWorldData {
     struct ExistingSpawn {
         ExistingSpawn(SpawnerTypeId id,
             const std::string& thing,
-            float density) :
+            double density) :
             id(id),
             thing(thing), setDensity(density)
         {}
@@ -21,7 +21,8 @@ class PatchManager : public Leviathan::PerWorldData {
         //! Used to detect existing items based on the chunk type or species
         //! name
         std::string thing;
-        float setDensity;
+        //! This is a double because the compound densities are a double
+        double setDensity;
 
         //! Flag for deleting removed ones. True when this should not be deleted
         bool marked = true;
